+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # 2.x branch
 ## 2.25 branch
+### 2.25.4
+* added `AppValidator::allowEmptyStringOnEmptyField()` rule method;
+* added `AppValidator::firstLetterCapitalized()` rule method;
+* fixed a serious bug with `FormHelper` which prevented the `radio` and "input group" (`append-text` and `prepend-text`
+  options)template from being customized at run-time;
+* removed the base `View` class (which was introduced with version 2.25.0. Backtracking, sorry);
+* `AppValidator::validPassword()` method has been simplified;
+* the `FormHelper` now makes extensive use of the `@inheritDoc` tag;
+* removed old (ancient) bake templates.
+
 ### 2.25.3
 * added `HtmlHelper::para()` method, so it can have icons;
 * fixed `AppTable::findActive()` method, now  it always uses the table alias so as not to generate conflicts;